Add Google HTTP Client docs, and rename googleclient module / package to googlehttpclient
