+++ pom.xml
@@ -36,7 +36,7 @@
     <module>jaxrs</module>
     <module>jaxrs2</module>
     <module>okhttp</module>
-    <module>googleclient</module>
+    <module>googlehttpclient</module>
     <module>ribbon</module>
     <module>sax</module>
     <module>slf4j</module>
@@ -68,7 +68,7 @@
     <main.basedir>${project.basedir}</main.basedir>
 
     <okhttp3.version>3.6.0</okhttp3.version>
-    <googleclient.version>1.31.0</googleclient.version>
+    <googlehttpclient.version>1.31.0</googlehttpclient.version>
     <gson.version>2.5</gson.version>
     <slf4j.version>1.7.13</slf4j.version>
     <bouncy.version>1.60</bouncy.version>
@@ -291,7 +291,7 @@
       <dependency>
         <groupId>com.google.http-client</groupId>
         <artifactId>google-http-client</artifactId>
-        <version>${googleclient.version}</version>
+        <version>${googlehttpclient.version}</version>
       </dependency>
 
       <dependency>
